requirements.txt
amqp==2.5.1
asn1crypto==1.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
defusedxml==0.6.0
Flask==1.1.1
graphyte==1.6.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
jira==2.0.0
kombu==4.6.5
MarkupSafe==1.1.1
more-itertools==7.2.0
oauthlib==3.1.0
pbr==5.4.3
prometheus-client==0.0.19
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.3
python-gitlab==1.11.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
six==1.12.0
urllib3==1.25.6
uWSGI==2.0.15
vine==1.3.0
Werkzeug==0.16.0
zipp==0.6.0
